# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/forgot-password": {
      "post": {
        "operationId": "forgot-password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "Auth",
          "Auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "reset-password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "Auth",
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hostly",
    "description": "Hostly API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}